+++ App/Zotero/components/zotero-service.js
@@ -29,14 +29,20 @@
 
 const Cc = Components.classes;
 const Ci = Components.interfaces;
+const Cr = Components.results;
+const Cu = Components.utils;
 
 /** XPCOM files to be loaded for all modes **/
 const xpcomFilesAll = [
 'zotero',
+'intl',
+'prefs',
 'dataDirectory',
 'date',
 'debug',
 'error',
+'utilities',
+'utilities_internal',
 'file',
 'http',
 'mimeTypeHandler',
@@ -49,9 +55,7 @@ const xpcomFilesAll = [
 'translation/translate_firefox',
 'translation/translator',
 'translation/tlds',
-'utilities',
 'isbn',
-'utilities_internal',
 'utilities_translate'];
 
 
@@ -101,6 +105,8 @@ const xpcomFilesLocal = [
 'mime',
 'notifier',
 'openPDF',
+'progressQueue',
+'progressQueueDialog',
 'quickCopy',
 'recognizePDF',
 'report',
@@ -338,14 +344,6 @@ function ZoteroService() {
 			makeZoteroContext(false);
 			zContext.Zotero.init(zInitOptions).
 			catch(function (e) {
-				if (e === "ZOTERO_SHOULD_START_AS_CONNECTOR") {
-					// if Zotero should start as a connector, reload it
-					return zContext.Zotero.shutdown().
-					then(function () {
-						makeZoteroContext(true);
-						return zContext.Zotero.init(zInitOptions);
-					});
-				}
 				dump(e + "\n\n");
 				Components.utils.reportError(e);
 				if (!zContext.Zotero.startupError) {
@@ -456,9 +454,7 @@ var _isStandalone = null;
                            */
 function isStandalone() {
 	if (_isStandalone === null) {
-		var appInfo = Components.classes["@mozilla.org/xre/app-info;1"].
-		getService(Components.interfaces.nsIXULAppInfo);
-		_isStandalone = appInfo.ID === 'zotero@chnm.gmu.edu';
+		_isStandalone = Services.appinfo.ID === 'zotero@chnm.gmu.edu';
 	}
 	return _isStandalone;
 }